add enemies (attackers)
an enemy (attacker) is created each time a new path is calculated by
the user clicking on a canvas cell or by clicking on the “send enemies”
button.  The cells on the path are “way points”.  The enemy does not
change direction abruptly but gradually by the application of an
acceleration vector at right angles to its velocity vector.
Parameters are loose to prevent the enemy from overshooting the center
of a cell but it still happens sometimes and then the enemy starts
spiraling away.  A better way needs to be found.