increase the velocity and smooth the motion of the enemy as it traverses the path
Increase the enemy velocity from 1.5 to 3.0.  Consider the enemy to
have reached the target when it is within 4 radii of the target (center
point of the cell), thus allowing it to start in a new direction
sooner.  Rotate the enemy velocity vector by a tenth of the angle
between it and the target each time through rather than the whole angle
between.